создает презентер карточки фильма
