+++ src/presenter/films-list-presenter.js
@@ -1,20 +1,13 @@
 import FilmsListView from '../view/films-list-view';
 import FilmsContainerView from '../view/films-container-view';
 import FilmsListContainerView from '../view/films-list-container-view';
-import FilmsCardView from '../view/films-card-view';
 import ButtonShowMoreView from '../view/show-more-button-view';
-import FilmsPopupView from '../view/films-popup-view';
 import FilmsSortView from '../view/sort-view';
-import CommentView from '../view/comment-view';
 import FilmsListNoCardsView from '../view/films-list-no-cards-view';
 import { render, RenderPosition, remove } from '../render';
-import { getRandomInteger } from '../utility';
-import { generateComment } from '../generator-data';
+import FilmsPresenter from './films-presenter';
 
 const NUMBER_CARDS_PER_STEP = 5;
-const MAX_NUMBER_COMMENTS = 5;
-const comments = Array.from({ length: MAX_NUMBER_COMMENTS }, generateComment);
-
 
 export default class FilmsListPresenter {
   #filmsContainer = null;
@@ -25,7 +18,6 @@ export default class FilmsListPresenter {
   #filmsListNoCardsComponent = new FilmsListNoCardsView();
   #filmsListContainerComponent = new FilmsListContainerView();
   #showMoreButton = new ButtonShowMoreView();
-
 
   #filmsCards = [];
   #renderedCards = NUMBER_CARDS_PER_STEP;
@@ -51,48 +43,8 @@ export default class FilmsListPresenter {
   }
 
   #renderCard = (card) => {
-    const popup = new FilmsPopupView(card);
-    const filmsCard = new FilmsCardView(card);
-
-    const body = document.querySelector('body');
-
-    const closePopup = () => {
-      remove(popup);
-      body.classList.remove('hide-overflow');
-    };
-
-    const onEscKeyDown = (evt) => {
-      if (evt.key === 'Escape' || evt.key === 'Esc') {
-        evt.preventDefault();
-        closePopup();
-        document.removeEventListener('keydown', onEscKeyDown);
-      }
-    };
-
-    const openPopup = () => {
-      body.classList.add('hide-overflow');
-      const footer = document.querySelector('.footer');
-      render(footer, popup, RenderPosition.AFTEREND);
-
-      const commentContainer = document.querySelector('.film-details__comments-list');
-
-      for (let i = 0; i < getRandomInteger(0, MAX_NUMBER_COMMENTS); i++) {
-        render(commentContainer, new CommentView(comments[i]), RenderPosition.BEFOREEND);
-      }
-
-      document.addEventListener('keydown', onEscKeyDown);
-
-      popup.setClosePopupClickHandler(() => {
-        document.removeEventListener('keydown', onEscKeyDown);
-        closePopup();
-      });
-    };
-
-    filmsCard.setFilmClickHandler(() => {
-      openPopup();
-    });
-
-    render(this.#filmsListContainerComponent, filmsCard, RenderPosition.BEFOREEND);
+    const cardPresenter = new FilmsPresenter(this.#filmsListContainerComponent);
+    cardPresenter.init(card);
   }
 
   #renderCardsList = () => {
